+++ Questionable/Data/AetheryteData.cs
@@ -2,8 +2,11 @@
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Numerics;
+using System.Runtime.CompilerServices;
 using Dalamud.Plugin.Services;
+using ECommons;
 using Lumina.Excel.Sheets;
+using Questionable.Functions;
 using Questionable.Model.Common;
 namespace Questionable.Data;
 
@@ -13,7 +16,6 @@ internal sealed class AetheryteData
     {
         Dictionary<EAetheryteLocation, uint> territoryIds = [];
         Dictionary<EAetheryteLocation, ushort> aethernetGroups = [];
-
 
         void ConfigureAetheryte(EAetheryteLocation aetheryteLocation, uint territoryId,
             ushort aethernetGroup)
@@ -307,6 +309,17 @@ internal sealed class AetheryteData
     public ReadOnlyDictionary<EAetheryteLocation, ushort> AethernetGroups { get; }
     private IReadOnlyList<uint> TownTerritoryIds { get; }
 
+    public EAetheryteLocation? NearestAetheryteTo(uint territoryId, Vector3 position)
+    {
+        return TerritoryIds
+                .Where(item => item.Value == territoryId && 
+                               !item.Key.IsAethernetShard() &&
+                               AetheryteFunctions.IsAetheryteUnlocked((uint)item.Key, out var _))
+                .Select(item => item.Key)
+                .OrderBy(key => CalculateDistance(position, territoryId, key))
+                .FirstOrNull();
+    }
+
     public float CalculateDistance(Vector3 fromPosition, uint fromTerritoryType, EAetheryteLocation to)
     {
         if (!TerritoryIds.TryGetValue(to, out uint toTerritoryType) || fromTerritoryType != toTerritoryType)
@@ -341,4 +354,10 @@ internal sealed class AetheryteData
     public bool IsAirshipLanding(EAetheryteLocation aetheryte) => AirshipLandingLocations.ContainsKey(aetheryte);
 
     public bool IsGoldSaucerAetheryte(EAetheryteLocation aetheryte) => TerritoryIds[aetheryte] is 144 or 388;
+}
+
+internal static class AetheryteLocationExtensions
+{
+    [MethodImpl( MethodImplOptions.AggressiveInlining )]
+	internal static Vector3 Position(this EAetheryteLocation aetheryteLocation, AetheryteData aetheryteData) => aetheryteData.Locations[aetheryteLocation];
 }